+++ src/Components/JSONDetails.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import { useContext, useState } from 'react';
-import { View, Text, TouchableOpacity, ScrollView, StyleSheet } from 'react-native';
-import { Appbar } from 'react-native-paper';
+import { View, TouchableOpacity, ScrollView, StyleSheet } from 'react-native';
+import { Appbar, Title } from 'react-native-paper';
 import FeatherIcon from 'react-native-vector-icons/Feather';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
+import { Text } from './Text';
 
 // @ts-ignore
 import JSONTree from 'react-native-json-tree';
@@ -19,7 +20,43 @@ export const JSONDetails: React.FC<IProps> = props => {
   const theme = useContext(ThemeContext);
   const [viewJSON, setViewJSON] = useState<boolean>(false);
 
-  const _renderLabel = ([raw]: string) => <Text style={{ fontSize: 14, color: theme.base0D }}>{`${raw} : `}</Text>;
+  const _renderLabel = viewJSON
+    ? ([raw]: string) => <Text style={{ fontSize: 14, color: theme.base0D }}>{`${raw} : `}</Text>
+    : '';
+  let _content = null;
+  try {
+    _content = (
+      <View style={[styles.container, { position: 'relative' }]}>
+        <ScrollView horizontal style={{ backgroundColor: theme.secondaryColor }}>
+          <ScrollView nestedScrollEnabled style={styles.scrollview}>
+            {viewJSON ? (
+              <JSONTree
+                data={typeof props.data === 'string' ? JSON.parse(props.data) : props.data}
+                labelRenderer={_renderLabel}
+                theme={theme}
+                invertTheme={false}
+              />
+            ) : (
+              <Text style={{ color: theme.textColorOne }}>
+                {typeof props.data !== 'string' ? JSON.stringify(props.data, null, 2) : props.data}
+              </Text>
+            )}
+          </ScrollView>
+        </ScrollView>
+      </View>
+    );
+  } catch (error) {
+    _content = (
+      <View style={[styles.container, { position: 'relative', backgroundColor: theme.secondaryColor, padding: 16 }]}>
+        <Title style={{ color: theme.textColorOne, marginBottom: 16 }}>
+          Oups... An error occurs with this JSON. Are you sure it is correctly formated ?
+        </Title>
+        <Text style={{ color: theme.textColorOne }}>
+          {typeof props.data !== 'string' ? JSON.stringify(props.data, null, 2) : props.data}
+        </Text>
+      </View>
+    );
+  }
 
   return (
     <View style={styles.container}>
@@ -44,16 +81,7 @@ export const JSONDetails: React.FC<IProps> = props => {
           />
         </TouchableOpacity>
       </Appbar.Header>
-
-      <ScrollView horizontal style={{ backgroundColor: theme.secondaryColor }}>
-        <ScrollView nestedScrollEnabled style={styles.scrollview}>
-          {viewJSON ? (
-            <JSONTree data={JSON.parse(props.data)} labelRenderer={_renderLabel} theme={theme} invertTheme={false} />
-          ) : (
-            <Text style={{ color: theme.textColorOne }}>{props.data}</Text>
-          )}
-        </ScrollView>
-      </ScrollView>
+      {_content}
     </View>
   );
 };