+++ src/Components/Main.tsx
@@ -6,13 +6,14 @@ import { Appbar, Searchbar, ActivityIndicator } from 'react-native-paper';
 import Item, { ITEM_HEIGHT } from './Item';
 import FeatherIcon from 'react-native-vector-icons/Feather';
 import { Settings } from './Settings';
-import { ILog, SourceType, RequestMethod, EnumSourceType, EnumFilterType } from '../types';
+import { ILog, SourceType, RequestMethod, EnumSourceType, EnumFilterType, EnumStatus } from '../types';
 import RNRequest from '../Core/Objects/RNRequest';
 import ReduxAction from '../Core/Objects/ReduxAction';
 import NRequest from '../Core/Objects/NRequest';
 import ConnectionInfo from '../Core/Objects/ConnectionInfo';
 import { ThemeContext } from '../Theme';
-import { xlsxWriter, formatDatas, mergeArrays, compare } from '../Utils/helpers';
+import { xlsxWriter, formatDatas, mergeArrays, compare, getStatus } from '../Utils/helpers';
+import { Indicator } from './Indicator';
 
 export interface IProps {
   testId?: string;
@@ -25,6 +26,13 @@ export interface IProps {
   connections: ConnectionInfo[];
   clearAll: Function;
   maxRequests?: number;
+  showStats?: boolean;
+}
+
+interface IStats {
+  success: number;
+  warning: number;
+  failed: number;
 }
 
 export const Main = (props: IProps) => {
@@ -33,7 +41,12 @@ export const Main = (props: IProps) => {
   const [searchQuery, setSearchQuery] = React.useState<string>('');
   const [source, setSource] = React.useState<SourceType | EnumSourceType>(EnumSourceType.All);
   const [filter, setFilter] = React.useState<RequestMethod | EnumFilterType>(EnumFilterType.All);
-  const [settingsVisible, setSettingsVisible] = React.useState(false);
+  const [stats, setStats] = React.useState<IStats>({
+    success: 100,
+    warning: 0,
+    failed: 0,
+  });
+  const [settingsVisible, setSettingsVisible] = React.useState<boolean>(false);
   const [deleteVisible, setDeleteVisible] = React.useState<boolean>(false);
   const [loadingXLSX, setloadingXLSX] = React.useState<boolean>(false);
   const onChangeSearch = (query: string) => setSearchQuery(query);
@@ -90,6 +103,40 @@ export const Main = (props: IProps) => {
     if (loadingXLSX) onShare();
   }, [loadingXLSX]);
 
+  React.useEffect(() => {
+    if (!props.showStats) return;
+    let _global: number = 0;
+    let _success: number = 0;
+    let _warning: number = 0;
+    let _failed: number = 0;
+
+    requests.map(item => {
+      if (item instanceof RNRequest || item instanceof NRequest) {
+        switch (getStatus(item.status)) {
+          case EnumStatus.Success:
+            _success++;
+            break;
+          case EnumStatus.Warning:
+            _warning++;
+            break;
+          default:
+            _failed++;
+            break;
+        }
+      }
+    });
+
+    _global = _success + _warning + _failed;
+    if (_global === 0) return;
+    const result = {
+      success: (_success * 100) / _global,
+      warning: (_warning * 100) / _global,
+      failed: (_failed * 100) / _global,
+    };
+
+    setStats(result);
+  }, [requests]);
+
   const clearList = () => {
     props.clearAll();
     setRequests([]);
@@ -110,7 +157,6 @@ export const Main = (props: IProps) => {
     } catch (error) {
       // if user dismiss sharing
       setloadingXLSX(false);
-      console.error(error.message);
     }
   };
 
@@ -219,22 +265,29 @@ export const Main = (props: IProps) => {
       {settingsVisible ? (
         <Settings source={source} onSetSource={setSource} filter={filter} onSetFilter={setFilter} />
       ) : (
-        <FlatList
-          testID="itemsList"
-          maintainVisibleContentPosition={{
-            autoscrollToTopThreshold: 10,
-            minIndexForVisible: 1,
-          }}
-          getItemLayout={_getItemLayout}
-          style={{ backgroundColor: theme.secondaryColor }}
-          renderItem={_renderItems}
-          keyExtractor={_keyExtractor}
-          data={requests}
-          initialNumToRender={20}
-          maxToRenderPerBatch={30}
-          updateCellsBatchingPeriod={70}
-          removeClippedSubviews
-        />
+        <>
+          {source === EnumSourceType.Redux ? (
+            <View style={[styles.indicatorPlaceholder, { backgroundColor: theme.reduxColor }]} />
+          ) : (
+            props.showStats && <Indicator success={stats.success} warning={stats.warning} failed={stats.failed} />
+          )}
+          <FlatList
+            testID="itemsList"
+            maintainVisibleContentPosition={{
+              autoscrollToTopThreshold: 10,
+              minIndexForVisible: 1,
+            }}
+            getItemLayout={_getItemLayout}
+            style={{ backgroundColor: theme.secondaryColor }}
+            renderItem={_renderItems}
+            keyExtractor={_keyExtractor}
+            data={requests}
+            initialNumToRender={20}
+            maxToRenderPerBatch={30}
+            updateCellsBatchingPeriod={70}
+            removeClippedSubviews
+          />
+        </>
       )}
     </>
   );
@@ -269,5 +322,9 @@ const styles = StyleSheet.create({
     height: 56,
     width: 56,
     borderLeftWidth: 1,
+  },
+  indicatorPlaceholder: {
+    height: 6,
+    width: '100%',
   },
 });